fix "garbage" characters printed on Windows cmd (ANSI terminal escape codes are not handled)
